🚚 Move helper to util
